SOLR-15646: Add a non secure RSAKeyPair keygen mode for tests.
